+++ CMakeLists.txt
@@ -290,7 +290,6 @@ if(${SUBSURFACE_TARGET_EXECUTABLE} MATCHES "MobileExecutable")
 		mobile-widgets/qml/kirigami/src/libkirigami/platformtheme.cpp
 		mobile-widgets/qml/kirigami/src/libkirigami/tabletmodewatcher.cpp
 		subsurface-mobile-main.cpp
-		subsurface-mobile-helper.cpp
 		subsurface-helper.cpp
 		profile-widget/qmlprofile.cpp
 		map-widget/qmlmapwidgethelper.cpp
