+++ subsurface-mobile-main.cpp
@@ -115,3 +115,21 @@ int main(int argc, char **argv)
 	free_prefs();
 	return 0;
 }
+
+void set_non_bt_addresses() {
+#if defined(Q_OS_ANDROID)
+	connectionListModel.addAddress("FTDI");
+#elif defined(Q_OS_LINUX) // since this is in the else, it does NOT include Android
+	connectionListModel.addAddress("/dev/ttyS0");
+	connectionListModel.addAddress("/dev/ttyS1");
+	connectionListModel.addAddress("/dev/ttyS2");
+	connectionListModel.addAddress("/dev/ttyS3");
+	// this makes debugging so much easier - use the simulator
+	connectionListModel.addAddress("/tmp/ttyS1");
+#endif
+}
+
+bool haveFilesOnCommandLine()
+{
+	return false;
+}
